pricing validation tweaks
